refactor: domain 디렉토리 루트 수정
