+++ backend/src/main/java/edonymyeon/backend/cache/application/HotPostsRedisRepository.java
@@ -1,6 +1,6 @@
 package edonymyeon.backend.cache.application;
 
-import edonymyeon.backend.cache.application.domain.CachedHotPost;
+import edonymyeon.backend.cache.domain.CachedHotPost;
 import org.springframework.data.repository.CrudRepository;
 
 public interface HotPostsRedisRepository extends CrudRepository<CachedHotPost, String> {
